requirements.txt
celery==5.4.0
email_validator==2.2.0
fastapi==0.115.6
PyJWT==2.10.1
python-dotenv==1.0.1
redis==5.2.1
six==1.17.0
SQLAlchemy==2.0.37
uvicorn==0.34.0
